type fixes

+++ PlusClouds.Net/Models/Country.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace PlusCloudsNet.Models
 {
     public class Country
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
         public string Name { get; set; }
         public string Code { get; set; }
         public string PhoneCode { get; set; }
