+++ PlusClouds.Net/Request/Products/SimilarProductsRequest.cs
@@ -1,7 +1,9 @@
+using System;
+
 namespace PlusCloudsNet.Request.Products
 {
     public class SimilarProductsRequest : BaseProductRequest
     {
-        public int Id { get; set; }
+        public Int64 Id { get; set; }
     }
 }
